refactor: DirectionUtil 삭제 후 도메인으로 분배
