+++ src/main/java/chess/domain/position/Position.java
@@ -1,12 +1,12 @@
 package chess.domain.position;
 
-import static chess.domain.position.direction.DirectionUtil.isInVerticalRange;
-
+import chess.domain.position.direction.VerticalDirection;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class Position {
 
@@ -63,7 +63,8 @@ public class Position {
     }
 
     public boolean isInVerticalRangeAndSameXAxis(Position other, int range) {
-        return isInVerticalRange(this, other, range) && isSameXAxis(other);
+        VerticalDirection verticalDirection = new VerticalDirection();
+        return verticalDirection.isInVerticalRange(this, other, range) && isSameXAxis(other);
     }
 
     public boolean isFarFromMoreThanOne(Position other) {
@@ -71,6 +72,26 @@ public class Position {
         int yAxisDelta = Math.abs(other.yAxis.getValue() - this.yAxis.getValue());
 
         return xAxisDelta > 1 || yAxisDelta > 1;
+    }
+
+    public static List<Position> getPositionsSameDirectionDiagonalBetween(Position from, Position to) {
+        int xAxisDelta = from.getXAxis().getValue() - to.getXAxis().getValue();
+        int yAxisDelta = from.getYAxis().getValue() - to.getYAxis().getValue();
+        int time = Math.abs(xAxisDelta);
+
+        int xDirection = -(xAxisDelta / time);
+        int yDirection = -(yAxisDelta / time);
+
+        return IntStream.range(1, time)
+                .mapToObj(idx -> getPositionWith(from, xDirection, yDirection, idx))
+                .collect(Collectors.toList());
+    }
+
+    private static Position getPositionWith(Position from, int xDir, int yDir, int idx) {
+        XAxis xAxis1 = XAxis.getByValue(from.getXAxis().getValue() + xDir * idx);
+        YAxis yAxis1 = YAxis.getByValue(from.getYAxis().getValue() + yDir * idx);
+
+        return Position.from(xAxis1, yAxis1);
     }
 
     public XAxis getXAxis() {
