+++ src/main/java/chess/domain/position/direction/VerticalDirection.java
@@ -11,4 +11,8 @@ public class VerticalDirection implements Direction{
     public boolean isOnVertical(Position from, Position to) {
         return from.isSameXAxis(to);
     }
+
+    public static boolean isInVerticalRange(Position from, Position other, int range) {
+        return Math.abs(from.getYAxis().subtract(other.getYAxis())) <= range;
+    }
 }
