+++ src/test/java/chess/domain/position/direction/VerticalDirectionTest.java
@@ -1,8 +1,6 @@
 package chess.domain.position.direction;
 
-import static chess.domain.position.direction.DirectionUtil.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import chess.domain.position.Position;
 import chess.domain.position.XAxis;
@@ -20,7 +18,8 @@ class VerticalDirectionTest {
         Position position2 = Position.from(XAxis.C, YAxis.ONE);
 
         // when
-        boolean actual = isInVerticalRange(position1, position2, 2);
+        VerticalDirection verticalDirection = new VerticalDirection();
+        boolean actual = verticalDirection.isInVerticalRange(position1, position2, 2);
 
         // then
         assertThat(actual).isTrue();
